requirements.txt
cassandra-driver
kafka-python
keras
matplotlib
numpy
pandas
pyspark
scikit-learn
scipy
seaborn
sklearn
tensorflow